Move some methods from `Entities` to `AbstractEntity`.
