+++ server/src/test/java/org/spine3/server/entity/EntityShould.java
@@ -54,8 +54,8 @@ import static org.mockito.Mockito.verify;
 import static org.spine3.base.Identifiers.newUuid;
 import static org.spine3.protobuf.Timestamps2.getCurrentTime;
 import static org.spine3.protobuf.Values.newStringValue;
-import static org.spine3.server.entity.Entities.createEntity;
-import static org.spine3.server.entity.Entities.getConstructor;
+import static org.spine3.server.entity.AbstractEntity.createEntity;
+import static org.spine3.server.entity.AbstractEntity.getConstructor;
 import static org.spine3.test.TimeTests.currentTimeSeconds;
 
 /**
